// src/api/inventory/models/Equipment.ts
export interface Equipment {
    /**
     * 
     * @type {string}
     * @memberof Equipment
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof Equipment
     */
    type: string;
    /**
     * Unique inventory/asset number
     * @type {string}
     * @memberof Equipment
     */
    inventoryNumber: string;
    /**
     * 
     * @type {Date}
     * @memberof Equipment
     */
    warrantyExpiry?: Date;
    /**
     * Current location ID
     * @type {string}
     * @memberof Equipment
     */
    locationId: string;
    /**
     * 
     * @type {string}
     * @memberof Equipment
     */
    notes?: string;
    /**
     * 
     * @type {string}
     * @memberof Equipment
     */
    id: string;
    /**
     * 
     * @type {EquipmentStatus}
     * @memberof Equipment
     */
    status: EquipmentStatus;
    /**
     * 
     * @type {Location}
     * @memberof Equipment
     */
    location?: Location;
    /**
     * 
     * @type {number}
     * @memberof Equipment
     */
    readonly openServiceRequestCount?: number;
    /**
     * 
     * @type {Date}
     * @memberof Equipment
     */
    createdAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof Equipment
     */
    updatedAt: Date;
}