// src/api/inventory/models/EquipmentStatus.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Hospital Equipment & Service Management API
 * API for managing hospital equipment inventory, service requests, and location tracking.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * - `ACTIVE` – in normal use
 * - `IN_SERVICE` – undergoing maintenance/repair
 * - `DECOMMISSIONED` – retired from service
 * @export
 */
export const EquipmentStatus = {
    Active: 'ACTIVE',
    InService: 'IN_SERVICE',
    Decommissioned: 'DECOMMISSIONED'
} as const;
export type EquipmentStatus = typeof EquipmentStatus[keyof typeof EquipmentStatus];


export function EquipmentStatusFromJSON(json: any): EquipmentStatus {
    return EquipmentStatusFromJSONTyped(json, false);
}

export function EquipmentStatusFromJSONTyped(json: any, ignoreDiscriminator: boolean): EquipmentStatus {
    return json as EquipmentStatus;
}

export function EquipmentStatusToJSON(value?: EquipmentStatus | null): any {
    return value as any;
}